# docs/swagger.yaml
basePath: /
definitions:
...
  server.LoginParams:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
...
  server.SuccessResponse:
    properties:
      message:
        type: string
    type: object
...
info:
  contact: {}
  description: API for managing accounts, certificate requests, and certificate authorities.
  title: Notary API
  version: "1.0"
paths:
...
  /login:
    post:
      consumes:
      - application/json
      description: Authenticates a user with email and password and sets a session
        cookie.
      parameters:
      - description: Login payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/server.LoginParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              $ref: '#/definitions/server.SuccessResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login
      tags:
      - auth
  /logout:
    post:
      description: Clears the current session cookie.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              $ref: '#/definitions/server.SuccessResponse'
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Logout
      tags:
      - auth
...
schemes:
- https
securityDefinitions:
  cookieAuth:
    in: cookie
    name: user_token
    type: apiKey
swagger: "2.0"